refactor: Move tag, posttag package
