+++ src/main/java/com/cherryhouse/server/heart/HeartService.java
@@ -5,8 +5,8 @@ import com.cherryhouse.server.post.Post;
 import com.cherryhouse.server.post.PostService;
 import com.cherryhouse.server.post.image.ImageMapping;
 import com.cherryhouse.server.post.image.ImageService;
-import com.cherryhouse.server.posttag.PostTagMapping;
-import com.cherryhouse.server.tag.TagService;
+import com.cherryhouse.server.post.posttag.PostTagMapping;
+import com.cherryhouse.server.post.tag.TagService;
 import com.cherryhouse.server.user.User;
 import com.cherryhouse.server.user.UserService;
 import lombok.RequiredArgsConstructor;
