+++ src/main/java/com/cherryhouse/server/post/PostService.java
@@ -7,8 +7,8 @@ import com.cherryhouse.server.post.image.ImageMapping;
 import com.cherryhouse.server.post.image.ImageService;
 import com.cherryhouse.server.post.dto.PostRequest;
 import com.cherryhouse.server.post.dto.PostResponse;
-import com.cherryhouse.server.posttag.PostTagMapping;
-import com.cherryhouse.server.tag.TagService;
+import com.cherryhouse.server.post.posttag.PostTagMapping;
+import com.cherryhouse.server.post.tag.TagService;
 import com.cherryhouse.server.user.User;
 import com.cherryhouse.server.user.UserService;
 import lombok.RequiredArgsConstructor;
